Rename components and references to components to PascalCase
