+++ src/components/app/header/Header.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import logo from '../../../assets/images/logo.svg';
-import './app-header.scss';
+import './Header.scss';
 
 export default function AppHeader() {
   return (
