+++ src/components/app/header/index.js
@@ -1,1 +1,1 @@
-export { default } from './app-header';
+export { default } from './Header';
