+++ src/routes/home/index.js
@@ -1,1 +1,1 @@
-export { default } from './home';
+export { default } from './Home';
