+++ src/routes/portfolio/index.js
@@ -1,1 +1,1 @@
-export { default } from './portfolio';
+export { default } from './Portfolio';
