+++ src/routes/portfolio/portfolio.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from './node_modules/react';
 
 // Component for /portfolio
 export default function Portfolio() {
